WheelError: move to wheel.wheelfile
The error is used when an error occurs anywhere in the wheel package.
The core part of the package is the wheelfile module. Because its needed
there, and to simplify vendoring of wheelfile let's move the definition
to that file. This change won't affect behavior. Additionally, relative
imports are used throughout the cli for consistency.