Append a `Build` tag to the WHEEL file
Currently, a package which is installed through a wheel loses the
information related to the wheel is was installed from. This is because
the build number is only used in the file name and this is lost while
installing the wheel. `bdist_wheel` appends a build tag to the `WHEEL`
file when a build number is passed. Extending `pack` to behave similarly
helps bring the wheel cli on par with what bdist_wheel provides and
allows debugging of regressed wheels.