Fixed installation from sdist (#334)
Wheel cannot use pep517 yet because the installation process involves installing wheel because the setuptools build backend declares it as a dependency for building wheels. A future update of pip may solve this problem.

Setuptools_scm, on the other hand, requires wheel as a build time dependency so if somebody tries to install wheel using `--no-binary=:all:`, it also creates a circular dependency.

Fixes #332.