CI: Reconfigure build matrix for Python 3.11 and PyPy 3.9
Reconfigures the build matrix to include Python 3.11 (which is currently in beta) and PyPy 3.9. PyPy 3.8 is also added and 3.7 is removed.

All Python versions (3.7 to 3.11, and PyPy 3.8 and 3.9) are tested on Ubuntu, and the latest stable versions (currently Python 3.10 and PyPy 3.9) also on macOS and Windows.

A weekly scheduled run is also added to catch changes in dependencies and environments (each Monday at 06:00 UTC), and workflow_dispatch is added to be able to manually trigger the CI workflow.