Fixed the most critical type issues and added a MyPy check to CI (#473)
* Explicitly mark missing setuptools attributes.

Suppresses Mypy errors.  Mypy will note if these ever become type hinted.

* Fix inconsistent timestamp hint.

This could be `int` and was passed to a non `int` sub-function.

For now I choose to make this more strict.
Alternatively `int` can be added to all timestamp hints.

* Fix BaseClass type hint.

Mypy got confused with the order and thought this was a type error.

All that matters is the variable gets declared with
`Type[ctypes.Structure]` first.  Since the other two are subclasses.

* Fix types for reused names.

More names are used or names are annotated with a union of the types
they'll have.

Fixes a minor bug in calculate_macosx_platform_tag.
`len(problematic_files)` was being called on the string, not the list.

* Use vendored types for vendored function calls.

Otherwise the types won't match, Mypy doesn't see them as the same.

This means a third party can't mix these types with a non-vendored
version, but that wouldn't be type safe anyway.

* Fix type alias to use compatible types.

This style was too new for early versions of Python.
It can and should be fixed without resorting to a TYPE_CHECKING branch.

* Work better with packaging types.

Removed cast as the variable was already the correct type.

Cast strings to their types using the class for those types.
The code looks suspect, I'm not sure if the regular expression is correct.

TYPE_CHECKING is no longer used.

* Add ignore for requirement.url attribute.

* Cast ZipFile IO[bytes] to BinaryIO.

This seems to be common problem with ZipFile that it uses the incorrect
type here.  This cast seems like the best way to fix it.

* Fix WheelContentElement types to match what's used in code.

If this is for the public API then it might be better to use a NamedTuple.

* ZipInfo wants an exact tuple size.

* Assert that dist-info match is correct.

Mypy does not like unchecked matching.

* Fix return statements.

If a function returns values then it can't implicitly return None. PEP 8

Functions returning values must have returns at the ends of all
branches.

* Exclude vendored sources from Mypy type checking.

* Add ignores for ctypes class magic.

I can think of better ways to do this, but not any which wouldn't
add a dependency.  I don't know enough about ctypes to suggest an
alternative method.  The Python struct module might work, but it'd be
some effort to convert to it.

* Add workaround for a bad upstream type hint.