cli: Fix removing build tag with `wheel pack --build-number ""`
There's an existing test for this case, which verifies that the original
build tag is removed from the wheel filename and is *not* removed from
the WHEEL metadata.  Presumably this is not what was intended.

BUILD_NUM_RE assumed that it would only match at end of file, but there's
at least a newline in the way.  Fix it, and the test.