fix: Fix extracted files permissions
This commit fixes the extracted file permissions
when using .extract/.extract_files in addition
to when using the unpack cli command.

The fix is by overriding the _extract_member
method of zip file, and adding permissions fix
there.

The permissions fix was moved from the unpack to
the new method.